+++ webknossos/webknossos/client/_download_dataset.py
@@ -5,12 +5,13 @@ import numpy as np
 from rich.progress import track
 from upath import UPath
 
-from .. import LayerCategoryType
 from ..dataset import Dataset
-from ..dataset.layer.abstract_layer import _element_class_to_dtype_per_channel
-from ..dataset_properties import (
-    LayerProperties,
+from ..dataset.defaults import DEFAULT_CHUNK_SHAPE, DEFAULT_DATA_FORMAT
+from ..dataset.layer.abstract_layer import (
+    LayerCategoryType,
+    _element_class_to_dtype_per_channel,
 )
+from ..dataset_properties import DataFormat, LayerProperties
 from ..geometry import BoundingBox, Mag, Vec3Int
 from .api_client.models import ApiUnusableDataSource
 from .context import _get_context
@@ -26,12 +27,14 @@ _DOWNLOAD_CHUNK_SHAPE = Vec3Int(512, 512, 512)
 
 def download_dataset(
     dataset_id: str,
+    *,
     sharing_token: str | None = None,
     bbox: BoundingBox | None = None,
     layers: list[str] | None = None,
     mags: list[Mag] | None = None,
     path: UPath | str | None = None,
     exist_ok: bool = False,
+    data_format: DataFormat | None = None,
 ) -> Dataset:
     context = _get_context()
     api_client = context.api_client
@@ -78,6 +81,7 @@ def download_dataset(
         layer = dataset.add_layer(
             layer_name=layer_name,
             category=category,
+            data_format=data_format or DEFAULT_DATA_FORMAT,
             dtype_per_channel=dtype_per_channel,
             num_channels=num_channels,
             largest_segment_id=getattr(api_data_layer, "largest_segment_id", None),
@@ -98,8 +102,10 @@ def download_dataset(
             mag_view = layer.get_or_add_mag(
                 mag,
                 compress=True,
-                chunk_shape=Vec3Int.full(32),
-                shard_shape=_DOWNLOAD_CHUNK_SHAPE,
+                chunk_shape=DEFAULT_CHUNK_SHAPE,
+                shard_shape=_DOWNLOAD_CHUNK_SHAPE
+                if data_format != DataFormat.Zarr
+                else DEFAULT_CHUNK_SHAPE,
             )
             aligned_bbox = layer.bounding_box.align_with_mag(mag, ceil=True)
             download_chunk_shape_in_mag = _DOWNLOAD_CHUNK_SHAPE * mag.to_vec3_int()